refactor: Organizar vista de Dashboard en subdirectorio
Este commit finaliza la tarea de organización de archivos de vistas, moviendo la vista del Dashboard a su propio subdirectorio dedicado.

- El archivo `views/dashboard_view.php` ha sido movido a `views/dashboard/dashboard_view.php`.
- El controlador `controllers/dashboard_controller.php` ha sido actualizado para reflejar la nueva ruta del archivo de la vista.

Con este cambio, todas las vistas principales de la aplicación están ahora organizadas en sus respectivos subdirectorios, mejorando la estructura y mantenibilidad del proyecto.
